Added AuditableDto for user module and exception class.

+++ commons/src/main/java/com/commons/utility/PageableResponse.java
@@ -15,7 +15,7 @@ public class PageableResponse <T>{
 
     private int currentPage;
 
-    private int totalRecord;
+    private long totalRecord;
 
     List<T> data;
 }
